med_agents/hallucination_detection/requirements.txt
torch>=2.0.0
transformers>=4.30.0
peft>=0.4.0
numpy>=1.20.0
scikit-learn>=1.0.0
datasets>=2.0.0
accelerate>=0.20.0
bitsandbytes>=0.39.0
python-dotenv>=1.0.0
